move scripts back to the old location in python-can/can/*
