--- can/scripts/__init__.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-"""
-This module contains various scripts, like a logfile writer and a logfile player.
-
-.. note::
-    The scripts reside in here so they can be launched as modules. That makes them
-    more easily callable on different platforms. But they they can also be called
-    directly (mostly on Unix systems), because they are also installed as scripts
-    from ``python-can/scripts/*``.
-    See PR #370.
-
-"""
