+++ scripts/can_logger.py
@@ -2,12 +2,12 @@
 # coding: utf-8
 
 """
-See :mod:`can.scripts.logger`.
+See :mod:`can.logger`.
 """
 
 from __future__ import absolute_import
 
-from can.scripts.logger import main
+from can.logger import main
 
 
 if __name__ == "__main__":
