+++ scripts/can_player.py
@@ -2,12 +2,12 @@
 # coding: utf-8
 
 """
-See :mod:`can.scripts.player`.
+See :mod:`can.player`.
 """
 
 from __future__ import absolute_import
 
-from can.scripts.player import main
+from can.player import main
 
 
 if __name__ == "__main__":
