+++ setup.cfg
@@ -15,12 +15,10 @@ branch = False
 omit =
     # legacy code
     can/CAN.py
-    can/logger.py
-    can/player.py
 
 [coverage:report]
 # two digits after decimal point
-precision = 2
+precision = 3
 show_missing = True
 exclude_lines =
     # Have to re-enable the standard pragma, see https://coverage.readthedocs.io/en/coverage-4.5.1a/config.html#syntax
