+++ test/test_scripts.py
@@ -19,9 +19,11 @@ class CanScriptTest(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        # clean out the argument list
+        # clean up the argument list so the call to the main() functions
+        # in test_does_not_crash() succeeds
         sys.argv = sys.argv[:1]
 
+    #: this is override by the subclasses
     __test__ = False
 
     __metaclass__ = ABCMeta
@@ -29,7 +31,8 @@ class CanScriptTest(unittest.TestCase):
     #@unittest.skipUnless(IS_UNIX, "commands may only be available on unix")
     def test_do_commands_exist(self):
         """This test calls each scripts once and veifies that the help
-        can be read without any errors.
+        can be read without any other errors, like the script not being
+        found.
         """
         for command in self._commands():
             try:
@@ -38,6 +41,9 @@ class CanScriptTest(unittest.TestCase):
                 self.assertEqual(e.returncode, errno.EINVAL,
                     'Calling "{}" failed (exit code was {} and not EINVAL/22):\n{}'
                     .format(command, e.returncode, e.output))
+            else:
+                # this is also okay
+                pass
 
     def test_does_not_crash(self):
         # test import
@@ -50,10 +56,15 @@ class CanScriptTest(unittest.TestCase):
 
     @abstractmethod
     def _commands(self):
+        """Returns an Iterable of commands that should "succeed", meaning they exit
+        normally (exit code 0) or with the exit code for invalid arguments: EINVAL/22.
+        """
         pass
 
     @abstractmethod
     def _import(self):
+        """Returns the modue of the script that has a main() function.
+        """
         pass
 
 
@@ -65,7 +76,6 @@ class TestLoggerScript(CanScriptTest):
         return (
             "can_logger.py --help",
             "python -m can.logger --help",
-            "python -m can.scripts.logger --help"
         )
 
     def _import(self):
@@ -81,7 +91,6 @@ class TestPlayerScript(CanScriptTest):
         return (
             "can_player.py --help",
             "python -m can.player --help",
-            "python -m can.scripts.player --help"
         )
 
     def _import(self):
